Update Windows Profile Transfer
Removed some transfer items. 
Added a bit that allows you to run this script from any machine instead of just a local machine. 
Added a bit to make sure the user has a profile on both machines. 
Added a bit to make sure that both computers are online before transfer